# docs/swagger.yaml
openapi: 3.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/VLYASHCHENKO01/comments/1.0.0
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/VLYASHCHENKO01/reviews/1.0.0
  - description: SwaggerHub API Auto Mocking
    url: http://localhost:8080/api/v1
info:
  version: 1.0.0
  title: Rest api reviews
  description: REST API implementation of comments for a product
paths:
...
  /products/{product_id}/reviews/{review_id}:
    parameters:
      - name: product_id
        in: path
        required: true
        description: Product identifie
        schema:
          type: integer
          format: int32
      - name: review_id
        in: path
        required: true
        description: Review identifie
        schema:
          type: integer
          format: int32
    patch:
      summary: Method for partilly updating a review
      tags:
        - Reviews
      operationId: updateCommentById
      requestBody:
        required: true
        content:
          application/vnd.api+json:
            schema:
              $ref: "#/components/schemas/ReviewUpdateRequest"
      responses:
        '200':
          description: Review successfully updated
          content: {}
        'default':
          description: Unforeseen server error
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/JSONAPIError"
    post:
      summary: Rate a review
      tags:
        - Rating
      description: Allows a user to rate a review
      operationId: rateReview
      requestBody:
        required: true
        content:
          application/vnd.api+json:
            schema:
              type: object
              properties:
                rating:
                  type: integer
                  description: Rating assigned to review by user
                  minimum: 1
                  maximum: 5
              required:
                - rating
      responses:
        '200':
          description: Rating successfully assigned to the review
          content:
            application/vnd.api+json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Rating successfully assigned to the review."
        'default':
          description: Unforeseen server error or other client errors
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/JSONAPIError"
    delete:
      summary: Delete a review rating
      tags:
        - Rating
      description: Allows a user to delete a rating assigned to a review
      operationId: deleteReviewRating
      responses:
        '204':
          description: Review rating successfully deleted
        '404':
          description: Review not found
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/JSONAPIError"
        'default':
          description: Unforeseen server error or other client errors
          content:
            application/vnd.api+json:
              schema:
                $ref: "#/components/schemas/JSONAPIError"
components:
  schemas:
...
    ReviewUpdate:
      type: object
      properties:
        rating:
          type: number
          format: float
          description: Product rating on a scale of 1 to 5, with two decimal points precision.
        content:
          type: string
          description: Text of the review
...
    ReviewUpdateRequest:
      type: object
      properties:
        data:
          type: object
          properties:
            type:
              type: string
              example: "reviews"
            id:
              type: string
              example: "123e4567-e89b-12d3-a456-426614174000"
            attributes:
              $ref: "#/components/schemas/ReviewUpdate"
    JSONAPIError:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            type: object
            required: [status, title, detail]
            properties:
              id:
                type: string
                description: Unique error identifier
              status:
                type: string
                description: String representation of HTTP error status
              code:
                type: string
                description: Internal error code
              title:
                type: string
                description: Brief description of the error
              detail:
                type: string
                description: Detailed description of the error
              links:
                type: object
                properties:
                  about:
                    type: string
                    format: uri
                    description: Link with more information about the error